### 2024-11-12 #7449 @arfon
@ChristianMorland – - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.

While you're waiting, you might want to fix up your ORCIDs, these should be formatted without the URL (e.g., `0000-0002-6719-0190`). You can check if your paper compiles properly with `@editorialbot generate pdf`
