### 2025-03-24 #7449 @samhforbes
@anastassiavybornova Does this paper look like one you'd be comfortable editing?
