### 2025-04-02 9902500
add osqp to requirements.txt
