### 2025-04-02 4e56738
set cvxpy dependency to newest version (1.6.4)
