### 2025-04-03 5807a5a
test manual activation of the coverage report
