### 2025-04-05 #7449 @anastassiavybornova
Thanks for your suggestions @ChristianMorland !

To the potential reviewers - @hariharanragothaman @olunusib @zeyus @bbortey9 @BradyPlanden - would you be able to review this submission to JOSS? Respond below and feel free to ping me if questions arise. If this would be your first JOSS review, I am happy provide further context and guidance on the review process.
