### 2025-04-09 #7449 @franziska-schier
> Hello [@ChristianMorland](https://github.com/ChristianMorland)! I will be editing your submission.
>
> In the next couple of days, I will double-check the items on the list below, and aim to find at least 2 reviewers for your submission. You're very welcome to suggest up to 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their github handles are most useful to receive but **please don't use "@" to reference them** since it will prematurely ping them.
>
> **Software**
>
>     * [x]  ∃ software description[x]  ∃ OSI-approved license[x]  ∃ installation instructions[ ]  ∃ documentation[x]  ∃ tests
>
>
> **Paper**
>
>     * [x]  Paper compiles[x]  Paper length 250-1000 words
>
>
> **Application**
>
>     * [x]  Substantial scholarly effort[x]  Clear research application
>
>
> Any questions along the way --> just ping me here.

Hello [@anastassiavybornova](https://github.com/anastassiavybornova),
I just want to let you (and potential reviewers) know that in the meanwhile we published the model documentation. It is availble on [GitHub](https://github.com/TI-Forest-Sector-Modelling/TiMBA)
