### 2025-04-10 #8034 @editorialbot
[REVIEW]: TiMBA
**Submitting author:** @ChristianMorland (<a href="http://orcid.org/0000-0001-6600-570X">Christian Morland</a>)
**Repository:** https://github.com/TI-Forest-Sector-Modelling/TiMBA
**Branch with paper.md** (empty if default branch): main_for_publication
**Version:** v.1.0.1
