### 2025-04-10 #8034 @anastassiavybornova
@hariharanragothaman @zeyus thanks again for agreeing to review this submission!

instructions on next steps for you as reviewers (starting by generating the checklist) are already posted by the editorialbot in the comments above. ping me if you need help or clarifications. you can ping the submitting author @ChristianMorland if you have questions / comments along the way.

a particular detail to pay attention to is the extended documentation, which i understand (@franziska-schier, feel free to add/correct) was added after the submission to joss, so i have not seen it myself yet.

and lastly @ChristianMorland if applicable and relevant - you're welcome to tag your coauthors here.

we aim to complete the review process in six weeks from now, that is, by May 22nd. i'll be checking in on your review conversation from time to time.

talk soon everyone!
