### 2025-04-10 #8034 @franziska-schier
> [@hariharanragothaman](https://github.com/hariharanragothaman) [@zeyus](https://github.com/zeyus) thanks again for agreeing to review this submission!
>
> instructions on next steps for you as reviewers (starting by generating the checklist) are already posted by the editorialbot in the comments above. ping me if you need help or clarifications. you can ping the submitting author [@ChristianMorland](https://github.com/ChristianMorland) if you have questions / comments along the way.
>
> a particular detail to pay attention to is the extended documentation, which i understand ([@franziska-schier](https://github.com/franziska-schier), feel free to add/correct) was added after the submission to joss, so i have not seen it myself yet.
>
> and lastly [@ChristianMorland](https://github.com/ChristianMorland) if applicable and relevant - you're welcome to tag your coauthors here.
>
> we aim to complete the review process in six weeks from now, that is, by May 22nd. i'll be checking in on your review conversation from time to time.
>
> talk soon everyone!

Thanks again [anastassiavybornova](https://github.com/anastassiavybornova) for coordinating this! Yes, the newly added documentation can be directly approached on github [here](https://github.com/TI-Forest-Sector-Modelling/TiMBA/blob/main/documentation/documentation_TiMBA.pdf)

If you, [@hariharanragothaman](https://github.com/hariharanragothaman) [@zeyus](https://github.com/zeyus), have any questions / need for clarification / etc. ping  [@ChristianMorland](https://github.com/ChristianMorland) , [@tomkeH](https://github.com/tomkeH) , [@Julia-ta](https://github.com/Julia-ta) or me ([@franziska-schier](https://github.com/franziska-schier))

We are looking forward to your comments
