### 2025-06-01 TI-Forest-Sector-Modelling/TiMBA#68 @zeyus
Coverage run results in a failure
Running the provided steps in the documentation, `coverage run` completes with an error.

Environment details:

- macOS 15.5
- Apple M3 Pro (Apple silicon)
- Python 3.9.6 (venv)
- Installed with `pip install -e .`

Output (last section) of `coverage run`:
```text
                                    Summary
(CVXPY) Jun 01 01:02:42 PM: Problem status: optimal
(CVXPY) Jun 01 01:02:42 PM: Optimal value: 2.332e+09
(CVXPY) Jun 01 01:02:42 PM: Compilation took 2.783e-03 seconds
(CVXPY) Jun 01 01:02:42 PM: Solver (including time spent in interface) took 1.782e+01 seconds
TiMBA     : INFO       ===========================
TiMBA     : INFO       Optimization problem for period 1 solved
TiMBA     : INFO       ===========================
TiMBA     : INFO       Problem status: optimal
TiMBA     : INFO       Optimal objective: -2332082447.2501793
TiMBA     : INFO       Number of iterations needed: 24425
TiMBA     : INFO       Run time: 17.818646916 sec.
TiMBA     : INFO       ===========================
TiMBA     : INFO       Number of corrected shadow prices in Demand: 21
/Users/xxx/github_projects/TiMBA/venv/lib/python3.9/site-packages/pandas/core/dtypes/cast.py:1641: DeprecationWarning: np.find_common_type is deprecated.  Please use `np.result_type` or `np.promote_types`.
See https://numpy.org/devdocs/release/1.25.0-notes.html and the docs for more information.  (Deprecated NumPy 1.25)
  return np.find_common_type(types, [])
TiMBA     : INFO       Shadow world prices activated
/Users/xxx/github_projects/TiMBA/venv/lib/python3.9/site-packages/pandas/core/dtypes/cast.py:1641: DeprecationWarning: np.find_common_type is deprecated.  Please use `np.result_type` or `np.promote_types`.
See https://numpy.org/devdocs/release/1.25.0-notes.html and the docs for more information.  (Deprecated NumPy 1.25)
  return np.find_common_type(types, [])
TiMBA     : INFO       Number of corrected shadow prices in Supply: 41
TiMBA     : INFO       Trade balance for following commodities is not satisfied
TiMBA     : INFO       Material Balance for all regions & commodities is satisfied
TiMBA     : INFO       Global Material Balance is not satisfied: -3240.9870537503157
TiMBA     : INFO       Verification supply upper bound :
TiMBA     : INFO       No supply bounds are exceeded
TiMBA     : INFO       Export bounds are not satisfied for 7 commodities
TiMBA     : INFO       Import bounds are not satisfied for 24 commodities
TiMBA     : INFO       Extraction done.
/Users/xxx/github_projects/TiMBA/venv/lib/python3.9/site-packages/pandas/core/dtypes/cast.py:1641: DeprecationWarning: np.find_common_type is deprecated.  Please use `np.result_type` or `np.promote_types`.
See https://numpy.org/devdocs/release/1.25.0-notes.html and the docs for more information.  (Deprecated NumPy 1.25)
  return np.find_common_type(types, [])
/Users/xxx/github_projects/TiMBA/venv/lib/python3.9/site-packages/pandas/core/dtypes/cast.py:1641: DeprecationWarning: np.find_common_type is deprecated.  Please use `np.result_type` or `np.promote_types`.
See https://numpy.org/devdocs/release/1.25.0-notes.html and the docs for more information.  (Deprecated NumPy 1.25)
  return np.find_common_type(types, [])
F
FAIL: test_timba_run (test_TiMBA.TestTiMBAClass)
Traceback (most recent call last):
  File "/Users/xxx/github_projects/TiMBA/test/test_TiMBA.py", line 73, in test_timba_run
    self.assertTrue(
AssertionError: False is not true : Produced TiMBA results do not correspond to provided validation results

Ran 8 tests in 128.408s

FAILED (failures=1)
````

Output of `coverage report`:

```text
Name                                                            Stmts   Miss Branch BrPart  Cover   Missing
-----------------------------------------------------------------------------------------------------------
TiMBA/__init__.py                                                   0      0      0      0   100%
TiMBA/data_management/DataContainer.py                            289     55     10      1    80%   38, 58, 64, 87, 92, 97, 102, 107, 112, 117, 122, 127, 132, 137, 142, 147, 152, 157, 162, 167, 170, 204, 212, 220, 228, 236, 244, 252, 260, 268, 276, 284, 292, 300, 308, 316, 324, 332, 350, 358, 366, 374, 380-381, 385, 389, 397, 432-433, 440, 443, 446, 449, 452-453
TiMBA/data_management/DataManager.py                              400     91     84      3    77%   24-29, 43-51, 64, 68, 73, 82-85, 232-235, 358-363, 395-412, 446-466, 547-550, 571-572, 593-594, 680-681, 760-796
TiMBA/data_management/ParameterCollector.py                       135     17      0      0    87%   47, 63, 71, 79, 87, 95, 103, 111, 119, 127, 135, 143, 151, 155, 159, 167, 170
TiMBA/data_management/__init__.py                                   0      0      0      0   100%
TiMBA/data_validation/DataValidator.py                             19      0      0      0   100%
TiMBA/data_validation/__init__.py                                   0      0      0      0   100%
TiMBA/helpers/__init__.py                                           0      0      0      0   100%
TiMBA/helpers/utils.py                                            137     37     28      2    64%   38-39, 43-46, 52-53, 59-60, 118, 156, 162, 184-213
TiMBA/logic/__init__.py                                             0      0      0      0   100%
TiMBA/logic/model.py                                              655    141    142     26    75%   69, 75->81, 102-104, 194-195, 212, 219, 251, 288->199, 290-294, 410-412, 781-1023, 1027->1074, 1119, 1158-1214, 1297-1300, 1305-1308, 1315-1320, 1325, 1329-1330, 1352, 1435-1436, 1571-1586, 1614->1624, 1654, 1680-1689, 1702-1718, 1738, 1740->1769, 1770-1771
TiMBA/logic/model_helpers.py                                      429     36     64     12    89%   30-32, 55-57, 81-83, 99->103, 101-102, 231-236, 477-480, 485-488, 557-558, 862-863, 942-944, 952-959, 962->974, 1119
TiMBA/logic/tests.py                                               73     18     28      9    69%   30, 36, 63, 70-74, 95, 99, 125-128, 157-158, 174-175
TiMBA/parameters/Defines.py                                        58      0      0      0   100%
TiMBA/parameters/Domains.py                                        32      0      0      0   100%
TiMBA/parameters/REGEX_patterns.py                                  2      0      0      0   100%
TiMBA/parameters/__init__.py                                       67     43     16      0    29%   48-63, 72-78, 87-99, 111-125
TiMBA/parameters/domain_lists.py                                   10      0      0      0   100%
TiMBA/parameters/domain_specifiers/AbstractDomainSpecifier.py      46      9      0      0    80%   7, 18, 26, 34, 42, 50, 55, 60, 68
TiMBA/parameters/domain_specifiers/Demand.py                       13      0      0      0   100%
TiMBA/parameters/domain_specifiers/ExogChange.py                  107      0      0      0   100%
TiMBA/parameters/domain_specifiers/Forest.py                       22      0      0      0   100%
TiMBA/parameters/domain_specifiers/Manufacture.py                  32      0      0      0   100%
TiMBA/parameters/domain_specifiers/RecyclingS.py                   11      0      0      0   100%
TiMBA/parameters/domain_specifiers/Specification.py                18      0      0      0   100%
TiMBA/parameters/domain_specifiers/Supply.py                       19      0      0      0   100%
TiMBA/parameters/domain_specifiers/Transportation.py               40      0      0      0   100%
TiMBA/parameters/domain_specifiers/__init__.py                      0      0      0      0   100%
TiMBA/parameters/headers.py                                         2      0      0      0   100%
TiMBA/parameters/paths.py                                          18      0      0      0   100%
TiMBA/results_logging/FileBaseclass.py                             46     18     16      5    53%   25-26, 30, 32-43, 48->52, 51, 58-60
TiMBA/results_logging/ResultsWriter.py                             15      8      2      0    41%   27-29, 37-41
TiMBA/results_logging/__init__.py                                   0      0      0      0   100%
TiMBA/results_logging/base_logger.py                               27      1      6      3    88%   14, 17->20, 21->36
TiMBA/user_io/__init__.py                                           0      0      0      0   100%
TiMBA/user_io/default_parameters.py                                17      0      0      0   100%
-----------------------------------------------------------------------------------------------------------
TOTAL                                                            2739    474    396     61    79%
```

I'm mainly reporting this for reference, as part of the review here: https://github.com/openjournals/joss-reviews/issues/8034

It's possible this is some floating point calculation difference based on hardware differences, but I'm not knowledgeable about this specific model to determine if the differences are within an acceptable range. Please let me know if I can supply more information to help debug this issue.
