### 2025-07-11 #8034 @ChristianMorland
Dear @zeyus,

thank you for your valuable feedback on our submission. We have taken the following steps to address the issues you raised:

1. We updated the README and added a brief "Known Issues" section to highlight compatibility issues with macOS and Python 3.13.
2. We corrected the references as suggested.
3. We made the link to the image absolute, and we hope that this will now ensure the figure is displayed properly.

The changes we made can be found within the branch 68-coverage-run-results-in-a-failure_v2.
The updated README now includes the following section:

"…with Windows and Ubuntu OS. The functionality with Python versions 3.9, 3.10, 3.11 and 3.12 on the latest Ubuntu and Windows OS versions is continuously tested using GitHub Actions.
Known Issues
TiMBA currently has compatibility issues on macOS and with Python 3.13. Users may experience unexpected behavior during the optimization process on macOS, as well as package installation problems when using Python 3.13. We recommend using Python 3.9-3.12 on Ubuntu or Windows for best results, until full support for Python 3.13 and macOS is implemented."

We believe these changes address your concerns and improve the overall quality of our submission. We look forward to your further evaluation.
