### 2025-07-15 #8034 @ayodhyasovis
> [@tobijjah](https://github.com/tobijjah) [@ayodhyasovis](https://github.com/ayodhyasovis) [@praneethd7](https://github.com/praneethd7) [@potterzot](https://github.com/potterzot) [@konstantinstadler](https://github.com/konstantinstadler) would any of you be available to review this submission to JOSS? that would be awesome! if you need some further details or clarifications, just reply directly here pinging me.

Hey, I can. But I also can give you the feedback next week.
