### 2025-09-22 TI-Forest-Sector-Modelling/TiMBA#78 @potterzot
Installation fails due to exporting pandas errors
I have python 3.12.4 and am able to install pandas in the venv manually, but `pip install -e .` fails with this message:

    UPDATING build\lib.win-amd64-cpython-312\pandas/_version.py
    set build\lib.win-amd64-cpython-312\pandas/_version.py to '1.5.3'
    running build_ext
    building 'pandas._libs.algos' extension
    error: Microsoft Visual C++ 14.0 or greater is required. Get it with "Microsoft C++ Build Tools": https://visualstudio.microsoft.com/visual-cpp-build-tools/

Despite the message about missing Visual C++ >=14.0, I have it installed and am able to install `pandas` in the venv manually but still receive this error when attempting `pip install -e .`.

Note that this error does not come from installing `pandas`. In `pip install -e .` Instead it is part of the setuptools configuration of this package. There are multiple failures to import pandas modules that share similar messages as this:

          ############################
          # Package would be ignored #
          ############################
          Python recognizes 'pandas.tests.io.data.legacy_pickle' as an importable package[^1],
          but it is absent from setuptools' `packages` configuration.

          This leads to an ambiguous overall configuration. If you want to distribute this
          package, please make sure that 'pandas.tests.io.data.legacy_pickle' is explicitly added
          to the `packages` configuration field.

          Alternatively, you can also rely on setuptools' discovery methods
          (for example by using `find_namespace_packages(...)`/`find_namespace:`
          instead of `find_packages(...)`/`find:`).

          You can read more about "package discovery" on setuptools documentation page:

          - https://setuptools.pypa.io/en/latest/userguide/package_discovery.html

          If you don't want 'pandas.tests.io.data.legacy_pickle' to be distributed and are
          already explicitly excluding 'pandas.tests.io.data.legacy_pickle' via
          `find_namespace_packages(...)/find_namespace` or `find_packages(...)/find`,
          you can try to use `exclude_package_data`, or `include-package-data=False` in
          combination with a more fine grained `package-data` configuration.

          You can read more about "package data files" on setuptools documentation page:

          - https://setuptools.pypa.io/en/latest/userguide/datafiles.html

          [^1]: For Python, any directory (with suitable naming) can be imported,
                even if it does not contain any `.py` files.
                On the other hand, currently there is no concept of package data
                directory, all directories are treated like packages.
