### 2025-09-22 #8034 @potterzot
Thank you for the reminder. At this time I am unable to test the functionality because the installation process fails. I have created an [issue](https://github.com/TI-Forest-Sector-Modelling/TiMBA/issues/78).

Specific Paper Comments:
- The following links do not seem to work:
    - The demand for wood-based products (end products in [figure 1](https://github.com/TI-Forest-Sector-Modelling/TiMBA/blob/main/paper/timba_product_transformation.png))
    - [timba_product_transformation\label{fig:1}](https://github.com/TI-Forest-Sector-Modelling/TiMBA/blob/main/paper/..\Cimages\Ctimba_product_transformation.png) Figure 1: Product structure and processing flows in TiMBA.
    - The supply of roundwood and related primary products ([figure 1](https://github.com/TI-Forest-Sector-Modelling/TiMBA/blob/main/paper/timba_product_transformation.png))
- There are three figures in [images](https://github.com/TI-Forest-Sector-Modelling/TiMBA/tree/main/images) but only one figure is referenced in the paper. Is this intentional?
