### 2025-10-06 #8034 @tomkeH
Hi @potterzot,
Thanks for raising this issue. The problem appears to occur only with Python versions >3.11, likely because Pandas 1.5.3 (used in TiMBA) does not support Python 3.12. We initially set the compatibility to Python 3.12.6 because the setup runs successfully on our PCs, likely due to the installed C++ Build Tools, and because our CI/CD pipeline completed without issues in that environment.

Since upgrading to Pandas 2.3.2 would require extensive changes in TiMBA, the developer team decided to limit compatibility to Python <3.12 for now. In the longer term, we plan to upgrade the Pandas and NumPy versions used in TiMBA.
The necessary changes to the installation files and the README, documenting the compatibility downgrade to Python <3.12, have already been made in the issue-related branch.

Regarding the figures in the images folder, the two remaining images are used in the TiMBA documentation, which is supplementary to the JOSS publication.

All changes related to the reviews of @potterzot and @zeyus are now merged in main.
