### 2025-11-07 #8034 @ChristianMorland
@anastassiavybornova, thank you for your helpful comments! We have made the suggested updates: we removed the timba_product_transformation caption, updated the Panayotou reference, changed the acknowledgments to "valuable support," and adjusted the paper title to match the Zenodo project. We would like to keep the current author order as is.

Could you let us know if a new release will be necessary to include these changes?
