### 2025-11-17 #8034 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
